+++ docs/users/changelog.md
@@ -7,6 +7,7 @@ Note that there is currently no guarantee for a stable Markdown formatting style
 
 - Removed
   - `mdformat.codepoints.ASCII_WHITESPACE` (deprecated since 0.7.20)
+  - `mdformat.plugins.ParserExtensionInterface.add_cli_options` (deprecated since 0.7.19)
 - Fixed
   - Read UTF-8 from standard input on all systems.
     Thank you, [Christopher Prohm](https://github.com/chmp), for the PR.